invariants: reorganize code
We add a `buildtags` package which defines constants for the `race`,
`cgo`, `tracing` tags and we simplify the `invariants` package.
